patch deploy script
